+++ pkg/characters/library/sam.go
@@ -2,8 +2,8 @@ package library
 
 func init() {
 	register(GenerateFromRegistry(CharacterMetadata{
-		Name:        "sa",
-		Description: "sa - Musical note avatar (Pure Red)",
+		Name:        "sam",
+		Description: "sam - Musical note avatar (Pure Red)",
 		Author:      "Wildreason, Inc",
 		Color:       "#FF0000",
 		Width:       11,